change state dtype to np.float32 for LavaCrossing and DoorKey

+++ src/academia/environments/door_key.py
@@ -105,4 +105,4 @@ class DoorKey(GenericMiniGridWrapper):
         elif self._door_status == 1 and 0 in door_array:
             self._door_status = 0
             
-        return np.array([*cells_flattened, direction, self._door_status])
+        return np.array([*cells_flattened, direction, self._door_status], dtype=np.float32)
